Add stub functions to db package
