+++ internal/db/repos_db_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"reflect"
 	"sort"
-	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -69,22 +68,6 @@ func mustCreate(ctx context.Context, t *testing.T, repos ...*types.Repo) []*type
 		createdRepos = append(createdRepos, repo)
 	}
 	return createdRepos
-}
-
-func generateRepos(n int, base ...*types.Repo) types.Repos {
-	if len(base) == 0 {
-		return nil
-	}
-
-	rs := make(types.Repos, 0, n)
-	for i := 0; i < n; i++ {
-		id := strconv.Itoa(i)
-		r := base[i%len(base)].Clone()
-		r.Name += api.RepoName(id)
-		r.ExternalRepo.ID += id
-		rs = append(rs, r)
-	}
-	return rs
 }
 
 // InsertRepoOp represents an operation to insert a repository.
